+++ wallet-spa/src/App.jsx
@@ -1,11 +1,7 @@
 import "./App.css";
 import React from "react";
-import { Route } from "react-router";
-import Home from "./pages/Home";
-import { AccountUI } from "./components/Account";
-import { AuthCallback } from "./pages/AuthCallback";
 import { Component } from "react";
-import { Container } from 'reactstrap';
+import Layout from "./components/Layout";
 
 export default class App extends Component {
   static displayName = App.name;
@@ -16,29 +12,7 @@ export default class App extends Component {
 
   render() {
     return (
-      <Container>
-        <Route
-          exact
-          path="/"
-          render={() => {
-            return <Home openIdManager={this.props.openIdManager} />;
-          }}
-        />
-        <Route
-          exact
-          path="/accounts/:accountId"
-          render={() => {
-            return <AccountUI />;
-          }}
-        />
-        <Route
-          path="/callback"
-          render={() => {
-            return <AuthCallback />;
-          }}
-        />
-
-      </Container>
+      <Layout openIdManager={this.props.openIdManager}/>
     )
   };
 }
