+++ wallet-spa/src/components/TopMenu.js
@@ -1,15 +1,15 @@
-import React from 'react';
-import { makeStyles } from '@material-ui/core/styles';
-import AppBar from '@material-ui/core/AppBar';
-import Toolbar from '@material-ui/core/Toolbar';
-import Typography from '@material-ui/core/Typography';
-import IconButton from '@material-ui/core/IconButton';
-import MenuIcon from '@material-ui/icons/Face';
-import { MenuItem } from '@material-ui/core';
+import React from "react";
+import { makeStyles } from "@material-ui/core/styles";
+import AppBar from "@material-ui/core/AppBar";
+import Toolbar from "@material-ui/core/Toolbar";
+import Typography from "@material-ui/core/Typography";
+import IconButton from "@material-ui/core/IconButton";
+import MenuIcon from "@material-ui/icons/Face";
+import { MenuItem } from "@material-ui/core";
 
 const drawerWidth = 240;
 
-const useStyles = makeStyles(theme => ({
+const useStyles = makeStyles((theme) => ({
   appBar: {
     width: `calc(100% - ${drawerWidth}px)`,
     marginLeft: drawerWidth,
@@ -27,24 +27,21 @@ function TopMenu() {
   const classes = useStyles();
 
   return (
-    <AppBar position='fixed' className={classes.appBar}>
+    <AppBar position="fixed" className={classes.appBar}>
       <Toolbar>
         <IconButton
-          edge='start'
+          edge="start"
           className={classes.menuButton}
-          color='inherit'
-          aria-label='menu'
+          color="inherit"
+          aria-label="menu"
         >
           <MenuIcon />
         </IconButton>
         <MenuItem>
-          <Typography variant='h6' className={classes.title}>
-            Home
-          </Typography>
-        </MenuItem>
-        <MenuItem>
-          <Typography variant='h6' className={classes.title}>
-            About
+          <Typography variant="h6" className={classes.title}>
+            <a className="text-dark" href="/">
+              Home
+            </a>
           </Typography>
         </MenuItem>
       </Toolbar>
