+++ wallet-spa/src/pages/OpenAccount.js
